+++ index.js
@@ -1,3 +1,3 @@
 export { default as label } from './src/Label';
 export { default as BitMap } from './src/BitMap';
-export { labelWidth } from './src/Common';
+export { labelWidth } from './src/LabelPlacers/Common';
