+++ src/LabelPlacers/Common.js
