+++ src/LabelPlacers/LabelPlacer.js
